Break access control into several files for ease of updating
