+++ hs_access_control/models/user.py
@@ -1,1254 +1,20 @@
-"""
-Access control classes for hydroshare.
-
-This module implements access control for hydroshare.  Privilege models act on users,
-groups, and resources, and determine what objects are accessible to which users.  These models:
-
-* determine three kinds of privilege
-
-   1) user membership in and privilege over groups.
-   2) user privilege over resources.
-   3) group privilege over resources.
-
-* track the provenance of privilege granting to:
-
-   1) allow accounting for what happened.
-   2) allow limited undo of prior actions.
-   3) limit each user to accumulating one privilege over a thing
-
-Notes and quandaries
---------------------
-
-* A resource or group can become unowned as a result of a user becoming inactive.
-  In this case, logic must not break.
-* users of ResourceAccess.edit_resources, ResourceAccess.edit_users, ResourceAccess.view_users
-  should be aware that
-
-    * Listed resources include all accessible resources (via user or group sharing)
-    * privileges are cumulative over all flags
-    * Thus, these are not lists for display; they are access lists.
-
-* In general, use get_resources_with_explicit_access to create display lists.
-
-* In general, the system reports "effective" privilege.
-
-    * effective privilege: that after accounting for resource flags (particularly, 'published'
-      and 'immutable'.
-    * declared privilege: that before accounting for resource flags.
-
-  Thus, it is possible that sharing something immutable with CHANGE privilege will result in
-  effective VIEW privilege.
-"""
-
 from django.contrib.auth.models import User, Group
 from django.db import models
-from django.db.models import Q, F, Max
-from django.db import transaction
+from django.db.models import Q
 from django.core.exceptions import PermissionDenied
 
 from hs_core.models import BaseResource
-
-######################################
-# Access control subsystem
-######################################
-
-# TODO: setting published flag prohibits user from changing resource flags???
-# TODO: setting published flag allows administrator to change resource flags???
-
-
-class PolymorphismError(Exception):
-    """ A function is called with an inappropriate combination of arguments """
-    # This is a generic exception.
-    pass
-
-
-class PrivilegeCodes(object):
-    """
-    Privilege codes describe what capabilities a user has for a thing
-    Privilege is a numeric code 1-4:
-
-        * 1 or PrivilegeCodes.OWNER:
-            the user owns the object.
-
-        * 2 or PrivilegeCodes.CHANGE:
-            the user can change the content of the object but not its state.
-
-        * 3 or PrivilegeCodes.VIEW:
-            the user can view but not change the object.
-
-        * 4 or PrivilegeCodes.NONE:
-            the user has no privilege over the object.
-    """
-    OWNER = 1
-    CHANGE = 2
-    VIEW = 3
-    NONE = 4
-    CHOICES = (
-        (OWNER, 'Owner'),
-        (CHANGE, 'Change'),
-        (VIEW, 'View')
-        # (NONE, 'None') : disallow "no privilege" lines
-    )
-    # Names of privileges for printing
-    NAMES = ('Unspecified', 'Owner', 'Change', 'View', 'None')
-
-
-class PrivilegeBase(models.Model):
-    """
-    Shared methods for Privilege handling
-
-    These polymorphic routines act on any type of keys, so that
-    they are independent of the types of the base class members.
-    They implement the major interactions with privileges, including
-    share, unshare, undo_share, and ensure that provenance and privilege
-    are always in sync with one another.
-    """
-    class Meta:
-        abstract = True
-
-    @classmethod
-    def get_privilege(cls, **kwargs):
-        """
-        Get an effective privilege for a pair of keys
-
-        This works for any pair of attributes that together, form a key.
-        Examples include:
-
-            * UserResourcePrivilege.get_privilege(user={X}, resource={Y})
-            * GroupResourcePrivilege.get_privilege(group={X}, resource={Y})
-            * UserGroupPrivilege.get_privilege(user={X}, group={Y})
-
-        **This is a system routine** and not recommended for use in application code.
-        """
-        try:
-            return cls.objects.get(**kwargs).privilege
-        except cls.DoesNotExist:
-            return PrivilegeCodes.NONE
-
-    @classmethod
-    def update(cls, **kwargs):
-        """
-        Update an effective privilege record without maintaining provenance.
-
-        This works for any pair of attributes that together, form a key, along with shared
-        parameters `grantor` and `privilege`.  Examples include:
-
-            * UserResourcePrivilege.update(user={X}, resource={Y}, privilege={Z}, grantor={W})
-            * GroupResourcePrivilege.update(group={X}, resource={Y}, privilege={Z}, grantor={W})
-            * UserGroupPrivilege.update(user={X}, group={Y}, privilege={Z}, grantor={W})
-
-        **This is a system routine** and not recommended for use in application code.
-        There are no access control rules applied; this routine is unconditional.
-        Only use this routine if you wish to completely bypass access control.
-        Note also that using this routine directly breaks provenance and disables undo.
-        """
-        grantor = kwargs['grantor']
-        privilege = kwargs.get('privilege', None)
-        if privilege is not None and privilege < PrivilegeCodes.NONE:
-            if 'privilege' in kwargs:
-                del kwargs['privilege']
-            del kwargs['grantor']
-            with transaction.atomic():
-                record, create = cls.objects.get_or_create(defaults={'privilege': privilege,
-                                                                     'grantor': grantor},
-                                                           **kwargs)
-                if not create:
-                    record.privilege = privilege
-                    record.grantor = grantor
-                    record.save()
-        else:
-            if 'privilege' in kwargs:
-                del kwargs['privilege']
-            del kwargs['grantor']
-            cls.objects.filter(**kwargs) \
-               .delete()
-
-    @classmethod
-    def share(cls, **kwargs):
-        """
-        Share an item with a user or group without maintaining provenance.
-
-        This works for any pair of attributes that together, form a key, along with shared
-        parameters `grantor` and `privilege`.  Examples include:
-
-            * UserResourcePrivilege.share(user={X}, resource={Y}, privilege={Z}, grantor={W})
-            * GroupResourcePrivilege.share(group={X}, resource={Y}, privilege={Z}, grantor={W})
-            * UserGroupPrivilege.share(user={X}, group={Y}, privilege={Z}, grantor={W})
-
-        A share operation is just an update operation.
-
-        **This is a system routine** and not recommended for use in application code.
-        There are no access control rules applied; this routine is unconditional.
-        Only use this routine if you wish to completely bypass access control.
-        Note also that using this routine directly breaks provenance and disables undo.
-        """
-        cls.update(cls, **kwargs)
-
-    @classmethod
-    def unshare(cls, **kwargs):
-        """
-        Unshare an item with a user or group without maintaining provenance.
-
-        This works for any pair of attributes that together, form a key, along with shared
-        parameter `grantor`.  Examples include:
-
-            * UserResourcePrivilege.share(user={X}, resource={Y}, grantor={W})
-            * GroupResourcePrivilege.share(group={X}, resource={Y}, grantor={W})
-            * UserGroupPrivilege.share(user={X}, group={Y}, grantor={W})
-
-        An unshare operation is just an update operation with privilege=PrivilegeCodes.NONE.
-
-        **This is a system routine** and not recommended for use in application code.
-        There are no access control rules applied; this routine is unconditional.
-        Only use this routine if you wish to completely bypass access control.
-        Note also that using this routine directly breaks provenance and disables undo.
-
-        An unshare operation is just an update with privilege NONE.
-        """
-        cls.update(cls, privilege=PrivilegeCodes.NONE, **kwargs)
-
-
-class UserGroupPrivilege(PrivilegeBase):
-    """ Privileges of a user over a group
-
-    Having any privilege over a group is synonymous with membership.
-
-    There is a reasonable meaning to PrivilegeCodes.NONE, which is to be
-    a group member without the ability to discover the identities of other
-    group members.  However, this is currently disallowed. It is used in the
-    provenance models to record removing a privilege.
-    """
-
-    privilege = models.IntegerField(choices=PrivilegeCodes.CHOICES,
-                                    editable=False,
-                                    default=PrivilegeCodes.VIEW)
-    start = models.DateTimeField(editable=False, auto_now=True)
-
-    user = models.ForeignKey(User,
-                             null=False,
-                             editable=False,
-                             related_name='u2ugp',
-                             help_text='user to be granted privilege')
-
-    group = models.ForeignKey(Group,
-                              null=False,
-                              editable=False,
-                              related_name='g2ugp',
-                              help_text='group to which privilege applies')
-
-    grantor = models.ForeignKey(User,
-                                null=False,
-                                editable=False,
-                                related_name='x2ugp',
-                                help_text='grantor of privilege')
-
-    class Meta:
-        unique_together = ('user', 'group')
-
-    def __str__(self):
-        """ Return printed depiction for debugging """
-        return str.format("<user '{}' (id={}) holds {} ({})" +
-                          " over group '{}' (id={})" +
-                          " via grantor '{}' (id={})>",
-                          str(self.user.username), str(self.user.id),
-                          PrivilegeCodes.NAMES[self.privilege],
-                          str(self.privilege),
-                          str(self.group.name), str(self.group.id),
-                          str(self.grantor.username), str(self.grantor.id))
-
-    @classmethod
-    def share(cls, **kwargs):
-        """
-        Share a group with a user and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param group: target group.
-        :param user: target user.
-        :param privilege: privilege 1-4.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            UserGroupPrivilege.share(group={X}, user={Y}, privilege={Z}, grantor={W}
-        """
-        if __debug__:
-            assert 'group' in kwargs
-            assert isinstance(kwargs['group'], Group)
-            assert 'user' in kwargs
-            assert isinstance(kwargs['user'], User)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-            assert 'privilege' in kwargs
-            assert \
-                kwargs['privilege'] >= PrivilegeCodes.OWNER and \
-                kwargs['privilege'] <= PrivilegeCodes.NONE
-            assert len(kwargs) == 4
-        cls.update(**kwargs)
-        UserGroupProvenance.update(**kwargs)
-
-    @classmethod
-    def unshare(cls, **kwargs):
-        """
-        Unshare a group with a user and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param group: target group.
-        :param user: target user.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            UserGroupPrivilege.unshare(group={X}, user={Y}, grantor={W})
-
-        Important: this does not guard against removing a single owner.
-
-        **This is a system routine** that should not be called directly by developers!
-        Use UserAccess.unshare_group_with_user instead. That routine avoids single-owner
-        deletion.
-        """
-        if __debug__:
-            assert 'group' in kwargs
-            assert isinstance(kwargs['group'], Group)
-            assert 'user' in kwargs
-            assert isinstance(kwargs['user'], User)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-            assert len(kwargs) == 3
-        cls.update(privilege=PrivilegeCodes.NONE, **kwargs)
-        UserGroupProvenance.update(privilege=PrivilegeCodes.NONE, **kwargs)
-
-    @classmethod
-    def undo_share(cls, **kwargs):
-        """
-        Undo a share a group with a user and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param group: target group.
-        :param user: target user.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            UserGroupPrivilege.undo_share(group={X}, user={Y}, grantor={W})
-
-        In practice:
-
-        The "undo" operation is independent of the privileges a user currently holds.
-        Suppose -- for example -- that a user holds CHANGE, grants that to another user,
-        and then loses CHANGE. The undo of the other user is still possible, even though the
-        original user no longer has the privilege.
-
-        Important: this does not guard against removing a single owner.
-
-        **This is a system routine** that should not be called directly by developers!
-        Use UserAccess.undo_share_group_with_user instead. That routine avoids single-owner
-        deletion.
-        """
-        if __debug__:
-            assert 'group' in kwargs
-            assert isinstance(kwargs['group'], Group)
-            assert 'user' in kwargs
-            assert isinstance(kwargs['user'], User)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-            assert len(kwargs) == 3
-        grantor = kwargs['grantor']
-        del kwargs['grantor']
-        # undo in provenance model; add a record that reinstates previous privilege.
-        UserGroupProvenance.undo_share(grantor=grantor, **kwargs)
-        # read that record and post to privilege table.
-        r = UserGroupProvenance.get_current_record(**kwargs)
-        cls.update(user=r.user, group=r.group, privilege=r.privilege, grantor=r.grantor)
-
-    @classmethod
-    def get_undo_users(cls, **kwargs):
-        """ Get a set of users for which a grantor can undo privilege
-
-        :param group: group to check
-        :param grantor: user that will undo privilege
-
-        Important: this does not guard against removing a single owner.
-
-        **This is a system routine** that should not be called directly by developers!
-        Use UserAccess.__get_group_undo_users instead. That routine avoids single-owner deletion.
-        """
-        if __debug__:
-            assert 'group' in kwargs
-            assert isinstance(kwargs['group'], Group)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-            assert len(kwargs) == 2
-        return UserGroupProvenance.get_undo_users(**kwargs)
-
-
-class UserResourcePrivilege(PrivilegeBase):
-    """ Privileges of a user over a resource
-
-    This model encodes privileges of individual users, like an access
-    control list; see GroupResourcePrivilege and UserGroupPrivilege
-    for other kinds of privilege.
-    """
-
-    privilege = models.IntegerField(choices=PrivilegeCodes.CHOICES,
-                                    editable=False,
-                                    default=PrivilegeCodes.VIEW)
-    start = models.DateTimeField(editable=False, auto_now=True)
-
-    user = models.ForeignKey(User,
-                             null=False,
-                             editable=False,
-                             related_name='u2urp',
-                             help_text='user to be granted privilege')
-
-    resource = models.ForeignKey(BaseResource,
-                                 null=False,
-                                 editable=False,
-                                 related_name='r2urp',
-                                 help_text='resource to which privilege applies')
-
-    grantor = models.ForeignKey(User,
-                                null=False,
-                                editable=False,
-                                related_name='x2urp',
-                                help_text='grantor of privilege')
-
-    class Meta:
-        unique_together = ('user', 'resource')
-
-    def __str__(self):
-        """ Return printed depiction for debugging """
-        return str.format("<user '{}' (id={}) holds {} ({})" +
-                          " over resource '{}' (id={})" +
-                          " via grantor '{}' (id={})>",
-                          str(self.user.username), str(self.user.id),
-                          PrivilegeCodes.NAMES[self.privilege],
-                          str(self.privilege),
-                          str(self.resource.title).encode('ascii'),
-                          str(self.resource.short_id).encode('ascii'),
-                          str(self.grantor.username), str(self.grantor.id))
-
-    @classmethod
-    def share(cls, **kwargs):
-        """
-        Share a resource with a user and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param resource: target resource.
-        :param user: target user.
-        :param privilege: privilege 1-4.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            UserResourcePrivilege.share(resource={X}, user={Y}, privilege={Z}, grantor={W})
-
-        This routine links UserResourcePrivilege and UserResourceProvenance.
-        """
-        cls.update(**kwargs)
-        UserResourceProvenance.update(**kwargs)
-
-    @classmethod
-    def unshare(cls, **kwargs):
-        """
-        Unshare a resource with a user and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param resource: target resource.
-        :param user: target user.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            UserResourcePrivilege.unshare(resource={X}, user={Y}, grantor={W})
-
-        This routine links UserResourcePrivilege and UserResourceProvenance.
-
-        Important: this does not guard against removing a single owner.
-
-        **This is a system routine** that should not be called directly by developers!
-        Use UserAccess.unshare_resource_with user instead. That routine avoids
-        single-owner deletion.
-        """
-        cls.update(privilege=PrivilegeCodes.NONE, **kwargs)
-        UserResourceProvenance.update(privilege=PrivilegeCodes.NONE, **kwargs)
-
-    @classmethod
-    def undo_share(cls, **kwargs):
-        """
-        Undo a share of a resource with a user and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param resource: target resource.
-        :param user: target user.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            UserResourcePrivilege.undo_share(resource={X}, user={Y}, grantor={W})
-
-        In practice:
-
-        The "undo" operation is independent of the privileges a user currently holds.
-        Suppose -- for example -- that a user holds CHANGE, grants that to another user,
-        and then loses CHANGE. The undo of the other user is still possible, even though the
-        original user no longer has the privilege.
-
-        This routine links UserResourcePrivilege and UserResourceProvenance.
-
-        Important: this does not guard against removing a single owner.
-
-        **This is a system routine** that should not be called directly by developers!
-        Use UserAccess.undo_share_resource_with user instead. That routine avoids
-        single-owner deletion.
-        """
-        if __debug__:
-            assert 'resource' in kwargs
-            assert isinstance(kwargs['resource'], BaseResource)
-            assert 'user' in kwargs
-            assert isinstance(kwargs['user'], User)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-        # add an undo record to the provenance table.
-        UserResourceProvenance.undo_share(**kwargs)
-        # read that record.
-        del kwargs['grantor']
-        r = UserResourceProvenance.get_current_record(**kwargs)
-        # post to privilege table.
-        cls.update(user=r.user, resource=r.resource, privilege=r.privilege, grantor=r.grantor)
-
-    @classmethod
-    def get_undo_users(cls, **kwargs):
-        """ Get a set of users for which the current user can undo privilege
-
-        :param resource: resource to check
-        :param grantor: user that will undo privilege
-
-        Important: this does not guard against removing a single owner.
-
-        **This is a system routine** that should not be called directly by developers!
-        Use UserAccess.__get_resource_undo_users instead. That routine avoids single-owner
-        deletion.
-        """
-        if __debug__:
-            assert 'resource' in kwargs
-            assert isinstance(kwargs['resource'], BaseResource)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-            assert len(kwargs) == 2
-        return UserResourceProvenance.get_undo_users(**kwargs)
-
-
-class GroupResourcePrivilege(PrivilegeBase):
-    """
-    Privileges of a group over a resource.
-
-    The group privilege over a resource is never directly meaningful;
-    it is resolved instead into user privilege for each member of
-    the group, as listed in UserGroupPrivilege above.
-    """
-
-    privilege = models.IntegerField(choices=PrivilegeCodes.CHOICES,
-                                    editable=False,
-                                    default=PrivilegeCodes.VIEW)
-    start = models.DateTimeField(editable=False, auto_now=True)
-
-    group = models.ForeignKey(Group,
-                              null=False,
-                              editable=False,
-                              related_name='g2grp',
-                              help_text='group to be granted privilege')
-
-    resource = models.ForeignKey(BaseResource,
-                                 null=False,
-                                 editable=False,
-                                 related_name='r2grp',
-                                 help_text='resource to which privilege applies')
-
-    grantor = models.ForeignKey(User,
-                                null=False,
-                                editable=False,
-                                related_name='x2grp',
-                                help_text='grantor of privilege')
-
-    class Meta:
-        unique_together = ('group', 'resource')
-
-    def __str__(self):
-        """ Return printed depiction for debugging """
-        return str.format("<group '{}' (id={}) holds {} ({})" +
-                          " over resource '{}' (id={})" +
-                          " via grantor '{}' (id={})>",
-                          str(self.group.name), str(self.group.id),
-                          PrivilegeCodes.NAMES[self.privilege],
-                          str(self.privilege),
-                          str(self.resource.title).encode('ascii'),
-                          str(self.resource.short_id).encode('ascii'),
-                          str(self.grantor.username), str(self.grantor.id))
-
-    @classmethod
-    def share(cls, **kwargs):
-        """
-        Share a resource with a group and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param resource: target resource.
-        :param group: target group.
-        :param privilege: privilege 1-4.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            GroupResourcePrivilege.share(resource={X}, group={Y}, privilege={Z}, grantor={W})
-
-        This routine links GroupResourceProvenance and GroupResourcePrivilege.
-        """
-        cls.update(**kwargs)
-        GroupResourceProvenance.update(**kwargs)
-
-    @classmethod
-    def unshare(cls, **kwargs):
-        """
-        Unshare a resource with a group and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param resource: target resource.
-        :param group: target group.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            GroupResourcePrivilege.unshare(resource={X}, group={Y}, grantor={W})
-
-        This routine links GroupResourceProvenance and GroupResourcePrivilege.
-        """
-        cls.update(privilege=PrivilegeCodes.NONE, **kwargs)
-        GroupResourceProvenance.update(privilege=PrivilegeCodes.NONE, **kwargs)
-
-    @classmethod
-    def undo_share(cls, **kwargs):
-        """
-        Undo a share of a resource with a group and update provenance
-
-        ***This completely bypasses access control*** but keeps provenance in sync.
-
-        :param resource: target resource.
-        :param group: target group.
-        :param grantor: user who requested privilege.
-
-        Usage:
-            GroupResourcePrivilege.undo_share(resource={X}, group={Y}, grantor={W})
-
-        In practice:
-
-        The "undo" operation is independent of the privileges a user currently holds.
-        Suppose -- for example -- that a user holds CHANGE, grants that to another user,
-        and then loses CHANGE. The undo of the other user is still possible, even though the
-        original user no longer has the privilege.
-
-        This routine links GroupResourceProvenance and GroupResourcePrivilege.
-        """
-        if __debug__:
-            assert 'resource' in kwargs
-            assert isinstance(kwargs['resource'], BaseResource)
-            assert 'group' in kwargs
-            assert isinstance(kwargs['group'], Group)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-        GroupResourceProvenance.undo_share(**kwargs)
-        del kwargs['grantor']
-        r = GroupResourceProvenance.get_current_record(**kwargs)
-        cls.update(group=r.group, resource=r.resource, privilege=r.privilege, grantor=r.grantor)
-
-    @classmethod
-    def get_undo_groups(cls, **kwargs):
-        """ Get a set of groups for which the current user can undo privilege
-
-        :param resource: resource to check
-        :param grantor: user that will undo privilege
-
-        **This is a system routine** that should not be called directly by developers!
-        Use UserAccess.__get_resource_undo_groups instead.
-        """
-        if __debug__:
-            assert 'resource' in kwargs
-            assert isinstance(kwargs['resource'], BaseResource)
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-            assert len(kwargs) == 2
-        return GroupResourceProvenance.get_undo_groups(**kwargs)
-
-
-class ProvenanceBase(models.Model):
-    """Methods reused by all provenance classes
-
-    These methods are independent of the types of the foreign keys they act upon,
-    and so can be reused in all provenance classes.
-
-    **Using any of these routines directly will break provenance.** Instead, they are
-    called from the specific "provenance" classes when needed.
-    """
-
-    class Meta:
-        abstract = True
-
-    def __str__(self):
-        """print name for debugging"""
-        return 'grantee=' + str(self.grantee) + \
-               ', entity=' + str(self.entity) + \
-               ', privilege=' + PrivilegeCodes.NAMES[self.privilege] + \
-               ' (' + str(self.privilege) + ')' + \
-               ', grantor=' + str(self.grantor) + \
-               ', start=' + str(self.start) + \
-               ', undone=' + str(self.undone)
-
-    @classmethod
-    def __get_current_start(cls, **kwargs):
-        """
-        Get the last start time for a given pair of keys.
-
-        This is the timestamp of the record that represents current state.
-
-        Usage:
-            UserResourceProvenance.__get_current_start(resource={X}, user={Y})
-            UserGroupProvenance.__get_current_start(group={X}, user={Y})
-            GroupResourceProvenance.__get_current_start(resource={X}, group={Y})
-        """
-        result = cls.objects\
-            .filter(**kwargs)\
-            .aggregate(Max('start'))
-        # This can be None if there is no start pair
-        return result['start__max']
-
-    @classmethod
-    def get_current_record(cls, **kwargs):
-        """
-        Get the current provenance record for a given pair of keys.
-
-        There is always an exact match between this record and the record in the
-        matching Privilege model.
-
-        Usage:
-            UserResourceProvenance.get_current_record(resource={X}, user={Y})
-            UserGroupProvenance.get_current_record(group={X}, user={Y})
-            GroupResourceProvenance.get_current_record(resource={X}, group={Y})
-        """
-        if __debug__:
-            assert len(kwargs) == 2
-        # First, compute the latest start time for the privilege.
-        # This is the start of the effective record.
-        start = cls.__get_current_start(**kwargs)
-        # Then, fetch that unique record.
-        if start is not None:
-            return cls.objects\
-                .get(start=start, **kwargs)
-        else:
-            return None
-
-    @classmethod
-    def get_privilege(cls, **kwargs):
-        """
-        Get the privilege implied by provenance state.
-
-        There is always an exact match between get_privilege for Provenance
-        and Privilege models.
-
-        Usage:
-            UserResourceProvenance.get_privilege(resource={X}, user={Y})
-            UserGroupProvenance.get_privilege(group={X}, user={Y})
-            GroupResourceProvenance.get_privilege(resource={X}, group={Y})
-        """
-        if __debug__:
-            assert len(kwargs) == 2
-        result = cls.get_current_record(**kwargs)
-        if result is not None:
-            return result.privilege
-        else:
-            return PrivilegeCodes.NONE
-
-    @classmethod
-    def update(cls, **kwargs):
-        """
-        Add a provenance record to the provenance chain.
-
-        The Provenance models are append-only tables, in the sense that no
-        record is ever deleted or changed. At any point in time the last record
-        entered for a pair is binding.  Records are automatically timestamped to enforce this.
-
-        Usage:
-            UserResourceProvenance.update(resource={X}, user={Y}, privilege={Z}, ...)
-            UserGroupProvenance.update(group={X}, user={Y}, privilege={Z}, ...)
-            GroupResourceProvenance.update(resource={X}, group={Y}, privilege={Z}, ...)
-        """
-        cls.objects.create(**kwargs)
-
-    @classmethod
-    def __get_prev_start(cls, **kwargs):
-        """
-        Get the previous start time.
-
-        An undo consists of backing up one time step from the current record to
-        find the previous one. This is then reinstated as the current record by
-        creating a copy. This can only be done once.
-
-        Usage:
-            UserResourceProvenance.__get_prev_start(resource={X}, user={Y})
-            UserGroupProvenance.__get_prev_start(group={X}, user={Y})
-            GroupResourceProvenance.__get_prev_start(resource={X}, group={Y})
-        """
-        if __debug__:
-            assert len(kwargs) == 2
-        last = cls.__get_current_start(**kwargs)
-        if last is not None:
-            result = cls.objects\
-                .filter(start__lt=last,
-                        **kwargs) \
-                .aggregate(Max('start'))
-            return result['start__max']
-        else:
-            return None
-
-    @classmethod
-    def __get_prev_record(cls, **kwargs):
-        """
-        Get the previous privilege record for a given pair.
-
-        This is the record that will be copied to reinstate its privilege after an undo.
-
-        Usage:
-            UserResourceProvenance.__get_prev_record(resource={X}, user={Y})
-            UserGroupProvenance.__get_prev_record(group={X}, user={Y})
-            GroupResourceProvenance.__get_prev_record(resource={X}, group={Y})
-        """
-        if __debug__:
-            assert len(kwargs) == 2
-        # First, compute the latest start time for the privilege.
-        # This is the start of the effective record.
-        start = cls.__get_prev_start(**kwargs)
-
-        # Then, fetch that (hopefully) unique record(s). There is a small chance of non-uniqueness.
-        if start is not None:
-            return cls.objects.get(start=start, **kwargs)
-        else:
-            return None
-
-    @classmethod
-    def undo_share(cls, **kwargs):
-        """
-        Undo one provenance change
-
-        To undo a change, one must find the last record and step back one time step.
-        One must also prevent backstep records from themselves being backstepped.
-
-        Usage:
-        ------
-
-                cls.undo_share({destination-args}, grantor={user})
-
-        where {destination-args} consist of the pair of entities (user/group, user/resource,
-        or group/resource) being controlled. The correct combination depends upon the subclass
-        in which this routine is inherited.  E.g.,
-
-                UserResourceProvenance.undo_share(resource={resource}, user={user}, grantor={user})
-
-        is one such combination.
-
-        Note that this does not modify the *Privilege class accordingly. That must be done
-        separately.
-
-        **This is never used directly in normal programming.** Use the routines in UserAccess.
-
-        Important: this does not protect against removing the last owner.
-        Supplementary logic is necessary
-        """
-        if __debug__:
-            assert 'grantor' in kwargs
-            assert isinstance(kwargs['grantor'], User)
-            assert len(kwargs) == 3
-        grantor = kwargs['grantor']
-        del kwargs['grantor']
-        current = cls.get_current_record(**kwargs)
-        if current is None:
-            raise PermissionDenied("No privilege to undo")
-        if current.undone:
-            raise PermissionDenied("Current privilege is already undone.")
-        if current.grantor != grantor:
-            raise PermissionDenied("Current user is not grantor")
-
-        previous = cls.__get_prev_record(**kwargs)
-        if previous is not None:
-            # create a rollback record that reinstates the previous privilege.
-            cls.update(privilege=previous.privilege,
-                       grantor=previous.grantor,
-                       undone=True,
-                       **kwargs)
-        else:
-            # put in a record revoking all privilege -- cannot be undone
-            cls.update(privilege=PrivilegeCodes.NONE,
-                       grantor=None,
-                       undone=True,
-                       **kwargs)
-
-    @classmethod
-    def share(cls, **kwargs):
-        """
-        Share a thing with a user or group
-
-        The thing can be a resource or group.  This is an unprotected method that does
-        an unconditional share.  Use this only to completely bypass access control.
-
-        Usage
-        -----
-
-                cls.share({destination-args}, privilege={privilege-code}, grantor={user})
-
-        where {destination-args} consist of the pair of entities (user/group, user/resource,
-        or group/resource) being controlled. The correct combination depends upon the subclass
-        in which this routine is inherited.  E.g.,
-
-                UserResourceProvenance.share(resource={resource}, user={user},
-                                            privilege={privilege}, grantor={user})
-
-        is one such combination.
-
-        Note that this does not modify the *Privilege class accordingly.
-        That must be done separately.
-        """
-        cls.update(**kwargs)
-
-    @classmethod
-    def unshare(cls, **kwargs):
-        """
-        Unshare a thing with a user or group
-
-        The thing can be a resource or group.  This is an unprotected method that does
-        an unconditional unshare.  Use this only for privileged unshare. Otherwise,
-        use undo_share
-
-        Usage
-        -----
-
-                cls.unshare({destination-args}, grantor={user})
-
-        where {destination-args} consist of the pair of entities (user/group, user/resource,
-        or group/resource) being controlled. this is exactly equivalent with:
-
-                cls.share(grantor={user}, {destination-args},
-                          privilege=PrivilegeCodes.NONE)
-
-        where {destination-args} consist of the pair of entities (user/group, user/resource,
-        or group/resource) being controlled. The correct combination depends upon the subclass
-        in which this routine is inherited.  E.g.,
-
-                UserResourceProvenance.unshare(resource={resource}, user={user}, grantor={user})
-
-        is one such combination.
-
-        Unshares can be undone using undo_share.
-
-        Note that this does not modify the *Privilege class accordingly.
-        That must be done separately.
-
-        """
-        cls.update(privilege=PrivilegeCodes.NONE, **kwargs)
-
-
-class UserGroupProvenance(ProvenanceBase):
-    """
-    Provenance of privileges of a user over a group
-
-    Having any privilege over a group is synonymous with membership.
-
-    This is an append-only ledger of group privilege that serves as complete provenance
-    of access changes.  At any time, one privilege applies to each grantee and group.
-    This is the privilege with the latest start date.  For performance reasons, this
-    information is cached in a separate table UserGroupPrivilege.
-
-    To undo a privilege, one appends a record to this table with PrivilegeCodes.NONE.
-    This is indistinguishable from having no record at all.  Thus, this provides a
-    complete time-based journal of what privilege was in effect when.
-
-    An "undone" field allows one-step undo but prohibits further undo.
-
-    """
-    privilege = models.IntegerField(choices=PrivilegeCodes.CHOICES,
-                                    editable=False,
-                                    default=PrivilegeCodes.VIEW)
-
-    start = models.DateTimeField(editable=False, auto_now_add=True)
-
-    user = models.ForeignKey(User,
-                             null=False,
-                             editable=False,
-                             related_name='u2ugq',
-                             help_text='user to be granted privilege')
-
-    group = models.ForeignKey(Group,
-                              null=False,
-                              editable=False,
-                              related_name='g2ugq',
-                              help_text='group to which privilege applies')
-
-    grantor = models.ForeignKey(User,
-                                null=True,
-                                editable=False,
-                                related_name='x2ugq',
-                                help_text='grantor of privilege')
-
-    undone = models.BooleanField(editable=False, default=False)
-
-    class Meta:
-        unique_together = ('user', 'group', 'start')
-
-    @property
-    def grantee(self):
-        """ make printing of privilege records work properly in superclass"""
-        return self.user
-
-    @property
-    def entity(self):
-        """ make printing of privilege records work properly in superclass"""
-        return self.group
-
-    @classmethod
-    def get_undo_users(cls, group, grantor):
-        """
-        get the users for which a specific grantee can undo privilege
-
-        :param group: group to check.
-        :param grantor: user that would initiate the rollback.
-
-        Note: undo is somewhat independent of access control. A user need not hold
-        a privilege to undo a privilege that was previously granted.
-        """
-
-        if __debug__:
-            assert isinstance(grantor, User)
-            assert isinstance(group, Group)
-
-        # users are those last granted a privilege over the entity by the grantor
-        # This syntax is curious due to undesirable semantics of .exclude.
-        # All conditions on the filter must be specified in the same filter statement.
-        selected = User.objects.filter(u2ugq__group=group)\
-                               .annotate(start=Max('u2ugq__start'))\
-                               .filter(u2ugq__start=F('start'),
-                                       u2ugq__grantor=grantor,
-                                       u2ugq__undone=False)
-        # launder out annotations used to select users
-        return User.objects.filter(pk__in=selected).exclude(pk=grantor.pk)
-
-    @classmethod
-    def update(cls, group, user, privilege, grantor, undone=False):
-        """
-        Add a provenance record to the provenance chain.
-
-        This is just a wrapper around ProvenanceBase.update that makes parameters explicit.
-        """
-
-        if __debug__:
-            assert isinstance(group, Group)
-            assert isinstance(user, User)
-            assert grantor is None or isinstance(grantor, User)
-            assert privilege >= PrivilegeCodes.OWNER and privilege <= PrivilegeCodes.NONE
-
-        super(UserGroupProvenance, cls).update(group=group,
-                                               user=user,
-                                               privilege=privilege,
-                                               grantor=grantor,
-                                               undone=undone)
-
-
-class UserResourceProvenance(ProvenanceBase):
-    """
-    Provenance of privileges of a user over a resource.
-
-    This is an append-only ledger of group privilege that serves as complete provenance
-    of access changes.  At any one time, one privilege applies to each user and group.
-    This is the privilege with the latest start date.  For performance reasons, this
-    information is cached in a separate table UserResourcePrivilege.
-
-    To undo a privilege, one appends a record to this table with PrivilegeCodes.NONE.
-    This is indistinguishable from having no record at all.  Thus, this provides a
-    time-based journal of what privilege was in effect when.
-
-    An "undone" field allows one-step undo but prohibits further undo.
-    """
-
-    privilege = models.IntegerField(choices=PrivilegeCodes.CHOICES,
-                                    editable=False,
-                                    default=PrivilegeCodes.VIEW)
-
-    start = models.DateTimeField(editable=False, auto_now_add=True)
-
-    user = models.ForeignKey(User,
-                             null=False,
-                             editable=False,
-                             related_name='u2urq',
-                             help_text='user to be granted privilege')
-
-    resource = models.ForeignKey(BaseResource,
-                                 null=False,
-                                 editable=False,
-                                 related_name='r2urq',
-                                 help_text='resource to which privilege applies')
-
-    grantor = models.ForeignKey(User,
-                                null=True,
-                                editable=False,
-                                related_name='x2urq',
-                                help_text='grantor of privilege')
-
-    undone = models.BooleanField(editable=False, default=False)
-
-    class Meta:
-        unique_together = ('user', 'resource', 'start')
-
-    @property
-    def grantee(self):
-        return self.user
-
-    @property
-    def entity(self):
-        return self.resource
-
-    @classmethod
-    def get_undo_users(cls, resource, grantor):
-        """ get the users for which a specific user can roll back privilege """
-
-        if __debug__:
-            assert isinstance(grantor, User)
-            assert isinstance(resource, BaseResource)
-
-        # users are those last granted a privilege over the resource by the grantor
-        # This syntax is curious due to undesirable semantics of .exclude.
-        # All conditions on the filter must be specified in the same filter statement.
-        selected = User.objects.filter(u2urq__resource=resource)\
-                               .annotate(start=Max('u2urq__start'))\
-                               .filter(u2urq__start=F('start'),
-                                       u2urq__grantor=grantor,
-                                       u2urq__undone=False)
-        # launder out annotations used to select users
-        return User.objects.filter(pk__in=selected).exclude(pk=grantor.pk)
-
-    @classmethod
-    def update(cls, resource, user, privilege, grantor, undone=False):
-        """
-        Add a provenance record to the provenance chain.
-
-        This is just a wrapper around ProvenanceBase.update with type checking.
-
-        The Provenance models are append-only tables, in the sense that no old records
-        are ever deleted; new records are added and timestamped as current.
-        """
-        if __debug__:
-            assert isinstance(resource, BaseResource)
-            assert isinstance(user, User)
-            assert grantor is None or isinstance(grantor, User)
-            assert privilege >= PrivilegeCodes.OWNER and privilege <= PrivilegeCodes.NONE
-
-        super(UserResourceProvenance, cls).update(resource=resource,
-                                                  user=user,
-                                                  privilege=privilege,
-                                                  grantor=grantor,
-                                                  undone=undone)
-
-
-class GroupResourceProvenance(ProvenanceBase):
-    """
-    Provenance of privileges of a group over a resource.
-
-    The group privilege over a resource is not directly meaningful.
-    it is resolved instead into user privilege for each member of
-    the group, as listed in UserGroupProvenance above.
-
-    This is an append-only ledger of group privilege that serves as complete provenance
-    of access changes.  At any one time, one privilege applies to each user and resource.
-    This is the privilege with the latest start date.  For performance reasons, this
-    information is cached in a separate table GroupResourcePrivilege.
-
-    To undo a privilege, one appends a record to this table with PrivilegeCodes.NONE.
-    This is indistinguishable from having no record at all.  Thus, this provides a
-    time-based journal of what privilege was in effect when.
-
-    An "undone" field allows one-step undo but prohibits further undo.
-
-    """
-
-    privilege = models.IntegerField(choices=PrivilegeCodes.CHOICES,
-                                    editable=False,
-                                    default=PrivilegeCodes.VIEW)
-
-    start = models.DateTimeField(editable=False, auto_now_add=True)
-
-    group = models.ForeignKey(Group,
-                              null=False,
-                              editable=False,
-                              related_name='g2grq',
-                              help_text='group to be granted privilege')
-
-    resource = models.ForeignKey(BaseResource,
-                                 null=False,
-                                 editable=False,
-                                 related_name='r2grq',
-                                 help_text='resource to which privilege applies')
-
-    grantor = models.ForeignKey(User,
-                                null=True,
-                                editable=False,
-                                related_name='x2grq',
-                                help_text='grantor of privilege')
-
-    undone = models.BooleanField(editable=False, default=False)
-
-    class Meta:
-        unique_together = ('group', 'resource', 'start')
-
-    @property
-    def grantee(self):
-        return self.group
-
-    @property
-    def entity(self):
-        return self.resource
-
-    @classmethod
-    def get_undo_groups(cls, resource, grantor):
-        """ get the groups for which a specific user can roll back privilege """
-
-        if __debug__:
-            assert isinstance(grantor, User)
-            assert isinstance(resource, BaseResource)
-
-        # groups are those last granted a privilege over the resource by the grantor
-        # This syntax is curious due to undesirable semantics of .exclude.
-        # All conditions on the filter must be specified in the same filter statement.
-        # We wish to avoid the state INITIAL, which cannot be undone.
-        # This is accomplished by setting a NULL grantor for INITIAL.
-        selected = Group.objects.filter(g2grq__resource=resource)\
-            .annotate(start=Max('g2grq__start'))\
-            .filter(g2grq__start=F('start'),
-                    g2grq__grantor=grantor,
-                    g2grq__undone=False)
-
-        # launder out annotations used to select users
-        return Group.objects.filter(pk__in=selected)
-
-    @classmethod
-    def update(cls, resource, group, privilege, grantor, undone=False):
-        """
-        Add a provenance record to the provenance chain.
-
-        This is just a wrapper around ProvenanceBase.update with argument type checking.
-        """
-        if __debug__:
-            assert isinstance(resource, BaseResource)
-            assert isinstance(group, Group)
-            assert grantor is None or isinstance(grantor, User)
-            assert privilege >= PrivilegeCodes.OWNER and privilege <= PrivilegeCodes.NONE
-        super(GroupResourceProvenance, cls).update(resource=resource,
-                                                   group=group,
-                                                   privilege=privilege,
-                                                   grantor=grantor,
-                                                   undone=undone)
+from hs_access_control.models.privilege import PrivilegeCodes, \
+        UserGroupPrivilege, UserResourcePrivilege, GroupResourcePrivilege
+from hs_access_control.models.group import GroupAccess, GroupMembershipRequest
+from hs_access_control.models.utils import PolymorphismError
+
+#############################################
+# Methods and data for users
+#
+# There is a one-one correspondence between isntances of UserAccess and instances of User.
+# UserAccess annotates each user with information and methods necessary for access control.
+#############################################
 
 
 class UserAccess(models.Model):
@@ -3477,514 +2243,3 @@ class UserAccess(models.Model):
             return self.can_undo_share_group_with_user(kwargs['group'], kwargs['user'])
         else:
             raise PolymorphismError(str.format("No action for arguments {}", str(kwargs)))
-
-
-class GroupMembershipRequest(models.Model):
-    request_from = models.ForeignKey(User, related_name='ru2gmrequest')
-
-    # when user is requesting to join a group this will be blank
-    # when a group owner is sending an invitation, this field will represent the inviting user
-    invitation_to = models.ForeignKey(User, null=True, blank=True, related_name='iu2gmrequest')
-    group_to_join = models.ForeignKey(Group, related_name='g2gmrequest')
-    date_requested = models.DateTimeField(editable=False, auto_now_add=True)
-
-
-class GroupAccess(models.Model):
-    """
-    GroupAccess is in essence a group profile object
-    Members are actually recorded in a separate model.
-    Membership is equivalent with holding some privilege over the group.
-    There is a well-defined notion of PrivilegeCodes.NONE for group,
-    which to be a member with no privileges over the group, including
-    even being able to view the member list. However, this is currently disallowed
-    """
-
-    # Django Group object: this has a side effect of creating Group.gaccess back relation.
-    group = models.OneToOneField(Group,
-                                 editable=False,
-                                 null=False,
-                                 related_name='gaccess',
-                                 related_query_name='gaccess',
-                                 help_text='group object that this object protects')
-
-    active = models.BooleanField(default=True,
-                                 editable=False,
-                                 help_text='whether group is currently active')
-
-    discoverable = models.BooleanField(default=True,
-                                       editable=False,
-                                       help_text='whether group description is discoverable' +
-                                                 ' by everyone')
-
-    public = models.BooleanField(default=True,
-                                 editable=False,
-                                 help_text='whether group members can be listed by everyone')
-
-    shareable = models.BooleanField(default=True,
-                                    editable=False,
-                                    help_text='whether group can be shared by non-owners')
-
-    auto_approve = models.BooleanField(default=False,
-                                       editable=False,
-                                       help_text='whether group membership can be auto approved')
-
-    description = models.TextField(null=False, blank=False)
-    purpose = models.TextField(null=True, blank=True)
-    date_created = models.DateTimeField(editable=False, auto_now_add=True)
-    picture = models.ImageField(upload_to='group', null=True, blank=True)
-
-    ####################################
-    # group membership: owners, edit_users, view_users are parallel to those in resources
-    ####################################
-
-    @property
-    def owners(self):
-        """
-        Return list of owners for a group.
-
-        :return: list of users
-
-        This eliminates duplicates due to multiple invitations.
-        """
-
-        return User.objects.filter(is_active=True,
-                                   u2ugp__group=self.group,
-                                   u2ugp__privilege=PrivilegeCodes.OWNER)
-
-    @property
-    def edit_users(self):
-        """
-        Return list of users who can add members to a group.
-
-        :return: list of users
-
-        This eliminates duplicates due to multiple invitations.
-        """
-
-        return User.objects.filter(is_active=True,
-                                   u2ugp__group=self.group,
-                                   u2ugp__privilege__lte=PrivilegeCodes.CHANGE)
-
-    @property
-    def members(self):
-        """
-        Return list of members for a group.
-
-        :return: list of users
-
-        This eliminates duplicates due to multiple invitations.
-        """
-
-        return User.objects.filter(is_active=True,
-                                   u2ugp__group=self.group,
-                                   u2ugp__privilege__lte=PrivilegeCodes.VIEW)
-
-    @property
-    def view_resources(self):
-        """
-        QuerySet of resources held by group.
-
-        :return: QuerySet of resource objects held by group.
-        """
-        return BaseResource.objects.filter(r2grp__group=self.group)
-
-    @property
-    def edit_resources(self):
-        """
-        QuerySet of resources that can be edited by group.
-
-        :return: List of resource objects that can be edited  by this group.
-        """
-        return BaseResource.objects.filter(r2grp__group=self.group, raccess__immutable=False,
-                                           r2grp__privilege__lte=PrivilegeCodes.CHANGE)
-
-    @property
-    def group_membership_requests(self):
-        """
-        get a list of pending group membership requests for this group (self)
-        :return: QuerySet
-        """
-
-        return GroupMembershipRequest.objects.filter(group_to_join=self.group,
-                                                     group_to_join__gaccess__active=True)
-
-    def get_resources_with_explicit_access(self, this_privilege):
-        """
-        Get a list of resources for which the group has the specified privilege
-
-        :param this_privilege: one of the PrivilegeCodes
-        :return: QuerySet of resource objects (QuerySet)
-
-        This routine is an attempt to organize resources for displayability. It looks at the
-        effective privilege rather than declared privilege, and squashes privilege that is in
-        conflict with resource flags. If the resource is immutable, it is reported as a "VIEW"
-        resource when the permission is "CHANGE", and as the original resource otherwise.
-        """
-        if __debug__:
-            assert this_privilege >= PrivilegeCodes.OWNER and this_privilege <= PrivilegeCodes.VIEW
-
-        # this query computes resources with privilege X as follows:
-        # a) There is a privilege of X for the object for group.
-        # b) There is no lower privilege in either group privileges for the object.
-        # c) Thus X is the effective privilege of the object.
-        if this_privilege == PrivilegeCodes.OWNER:
-            return BaseResource.objects.filter(r2grp__privilege=this_privilege,
-                                               r2grp__group=self.group)\
-                                       .exclude(pk__in=BaseResource.objects
-                                                .filter(r2grp__group=self.group,
-                                                        r2grp__privilege__lt=this_privilege))
-
-        elif this_privilege == PrivilegeCodes.CHANGE:
-            # CHANGE does not include immutable resources
-            return BaseResource.objects.filter(raccess__immutable=False,
-                                               r2grp__privilege=this_privilege,
-                                               r2grp__group=self.group)\
-                                       .exclude(pk__in=BaseResource.objects
-                                                .filter(r2grp__group=self.group,
-                                                        r2grp__privilege__lt=this_privilege))
-
-        else:  # this_privilege == PrivilegeCodes.ViEW
-            # VIEW includes CHANGE & immutable as well as explicit VIEW
-            view = BaseResource.objects.filter(r2grp__privilege=this_privilege,
-                                               r2grp__group=self.group)\
-                .exclude(pk__in=BaseResource.objects.filter(r2grp__group=self.group,
-                                                            r2grp__privilege__lt=this_privilege))
-
-            immutable = BaseResource.objects.filter(raccess__immutable=True,
-                                                    r2grp__privilege=PrivilegeCodes.CHANGE,
-                                                    r2grp__group=self.group)\
-                                            .exclude(pk__in=BaseResource.objects.filter(
-                                                raccess__immutable=True,
-                                                r2grp__group=self.group,
-                                                r2grp__privilege__lt=PrivilegeCodes.CHANGE))
-
-            return BaseResource.objects.filter(Q(pk__in=view) | Q(pk__in=immutable)).distinct()
-
-    def get_users_with_explicit_access(self, this_privilege):
-        """
-            Get a list of users for which the group has the specified privilege
-
-            :param this_privilege: one of the PrivilegeCodes
-            :return: QuerySet of user objects (QuerySet)
-        """
-
-        if __debug__:
-            assert this_privilege >= PrivilegeCodes.OWNER and this_privilege <= PrivilegeCodes.VIEW
-
-        if this_privilege == PrivilegeCodes.OWNER:
-            return self.owners
-        elif this_privilege == PrivilegeCodes.CHANGE:
-            return User.objects.filter(is_active=True,
-                                       u2ugp__group=self.group,
-                                       u2ugp__privilege=PrivilegeCodes.CHANGE)
-
-        else:  # this_privilege == PrivilegeCodes.VIEW
-            return User.objects.filter(is_active=True,
-                                       u2ugp__group=self.group,
-                                       u2ugp__privilege=PrivilegeCodes.VIEW)
-
-    def get_effective_privilege(self, this_user):
-        """
-        Return cumulative privilege for a user over a group
-
-        :param this_user: User to check
-        :return: Privilege code 1-4
-        """
-
-        if not this_user.is_active:
-            return PrivilegeCodes.NONE
-        try:
-            p = UserGroupPrivilege.objects.get(group=self.group,
-                                               user=this_user)
-            return p.privilege
-        except UserGroupPrivilege.DoesNotExist:
-            return PrivilegeCodes.NONE
-
-
-class ResourceAccess(models.Model):
-    """ Resource model for access control
-    """
-    #############################################
-    # model variables
-    #############################################
-
-    resource = models.OneToOneField(BaseResource,
-                                    editable=False,
-                                    null=False,
-                                    related_name='raccess',
-                                    related_query_name='raccess')
-
-    # only for resources
-    active = models.BooleanField(default=True,
-                                 help_text='whether resource is currently active')
-    # both resources and groups
-    discoverable = models.BooleanField(default=False,
-                                       help_text='whether resource is discoverable by everyone')
-    public = models.BooleanField(default=False,
-                                 help_text='whether resource data can be viewed by everyone')
-    shareable = models.BooleanField(default=True,
-                                    help_text='whether resource can be shared by non-owners')
-    # these are for resources only
-    published = models.BooleanField(default=False,
-                                    help_text='whether resource has been published')
-    immutable = models.BooleanField(default=False,
-                                    help_text='whether to prevent all changes to the resource')
-
-    require_download_agreement = models.BooleanField(default=False,
-                                                     help_text='whether to require agreement to '
-                                                               'resource rights statement for '
-                                                               'resource content downloads')
-    #############################################
-    # workalike queries adapt to old access control system
-    #############################################
-
-    @property
-    def view_users(self):
-        """
-        QuerySet of users with view privileges over a resource.
-
-        This is a property so that it is a workalike for a prior explicit list.
-
-        For VIEW, effective privilege = declared privilege, in the sense that all editors have
-        VIEW, even if the resource is immutable.
-        """
-
-        return User.objects.filter(Q(is_active=True) &
-                                   (Q(u2urp__resource=self.resource,
-                                      u2urp__privilege__lte=PrivilegeCodes.VIEW) |
-                                    Q(u2ugp__group__gaccess__active=True,
-                                      u2ugp__group__g2grp__resource=self.resource,
-                                      u2ugp__group__g2grp__privilege__lte=PrivilegeCodes.VIEW)))\
-                           .distinct()
-
-    @property
-    def edit_users(self):
-        """
-        QuerySet of users with change privileges
-
-        This is a property so that it is a workalike for a prior explicit list.
-
-        If the resource is immutable, an empty QuerySet is returned.
-        """
-
-        if self.immutable:
-            return User.objects.none()
-        else:
-            return User.objects\
-                       .filter(Q(is_active=True) &
-                               (Q(u2urp__resource=self.resource,
-                                  u2urp__privilege__lte=PrivilegeCodes.CHANGE) |
-                                Q(u2ugp__group__gaccess__active=True,
-                                  u2ugp__group__g2grp__resource=self.resource,
-                                  u2ugp__group__g2grp__privilege__lte=PrivilegeCodes.CHANGE)))\
-                .distinct()
-
-    @property
-    def view_groups(self):
-        """
-        QuerySet of groups with view privileges
-
-        This is a property so that it is a workalike for a prior explicit list
-        """
-        return Group.objects.filter(g2grp__resource=self.resource,
-                                    g2grp__privilege__lte=PrivilegeCodes.VIEW, gaccess__active=True)
-
-    @property
-    def edit_groups(self):
-        """
-        QuerySet of groups with edit privileges
-
-        This is a property so that it is a workalike for a prior explicit list.
-
-        If the resource is immutable, an empty QuerySet is returned.
-        """
-        if self.immutable:
-            return Group.objects.none()
-        else:
-            return Group.objects.filter(g2grp__resource=self.resource,
-                                        g2grp__privilege__lte=PrivilegeCodes.CHANGE,
-                                        gaccess__active=True)
-
-    @property
-    def owners(self):
-        """
-        QuerySet of users with owner privileges
-
-        This is a property so that it is a workalike for a prior explicit list.
-
-        For immutable resources, owners are not modified, but do not have CHANGE privilege.
-        """
-        return User.objects.filter(is_active=True,
-                                   u2urp__privilege=PrivilegeCodes.OWNER,
-                                   u2urp__resource=self.resource)
-
-    def get_users_with_explicit_access(self, this_privilege, include_user_granted_access=True,
-                                       include_group_granted_access=True):
-
-        """
-        Gets a QuerySet of Users who have the explicit specified privilege access to the resource.
-        An empty list is returned if both include_user_granted_access and
-        include_group_granted_access are set to False.
-
-        :param this_privilege: the explict privilege over the resource for which list of users
-         needed
-        :param include_user_granted_access: if True, then users who have been granted directly the
-        specified privilege will be included in the list
-        :param include_group_granted_access: if True, then users who have been granted the
-        specified privilege via group privilege over the resource will be included in the list
-        :return:
-        """
-        if include_user_granted_access and include_group_granted_access:
-            return User.objects.filter(Q(is_active=True) &
-                                       (Q(u2urp__resource=self.resource,
-                                          u2urp__privilege=this_privilege) |
-                                        Q(u2ugp__group__g2grp__resource=self.resource,
-                                          u2ugp__group__g2grp__privilege=this_privilege)))\
-                               .distinct()
-        elif include_user_granted_access:
-            return User.objects.filter(Q(is_active=True) &
-                                       (Q(u2urp__resource=self.resource,
-                                          u2urp__privilege=this_privilege)))
-        elif include_group_granted_access:
-            return User.objects.filter(Q(is_active=True) &
-                                       (Q(u2ugp__group__g2grp__resource=self.resource,
-                                          u2ugp__group__g2grp__privilege=this_privilege)))\
-                               .distinct()
-        else:
-            return []
-
-    def __get_raw_user_privilege(self, this_user):
-        """
-        Return the user-based privilege of a specific user over this resource
-
-        :param this_user: the user upon which to report
-        :return: integer privilege 1-4 (PrivilegeCodes)
-
-        This does not account for resource flags.
-
-        """
-        if __debug__:  # during testing only, check argument types and preconditions
-            assert isinstance(this_user, User)
-
-        if not this_user.is_active:
-            raise PermissionDenied("Grantee user is not active")
-
-        if this_user.is_superuser:
-            return PrivilegeCodes.OWNER
-
-        # compute simple user privilege over resource
-        try:
-            p = UserResourcePrivilege.objects.get(resource=self.resource,
-                                                  user=this_user)
-            response1 = p.privilege
-        except UserResourcePrivilege.DoesNotExist:
-            response1 = PrivilegeCodes.NONE
-
-        return response1
-
-    def __get_raw_group_privilege(self, this_user):
-        """
-        Return the group-based privilege of a specific user over this resource
-
-        :param this_user: the user upon which to report
-        :return: integer privilege 1-4 (PrivilegeCodes)
-
-        This does not account for resource flags.
-        """
-        if __debug__:  # during testing only, check argument types and preconditions
-            assert isinstance(this_user, User)
-
-        if not this_user.is_active:
-            raise PermissionDenied("Grantee user is not active")
-
-        if this_user.is_superuser:
-            return PrivilegeCodes.OWNER
-
-        # Group privileges must be aggregated
-        group_priv = GroupResourcePrivilege.objects\
-            .filter(resource=self.resource,
-                    group__gaccess__active=True,
-                    group__g2ugp__user=this_user)\
-            .aggregate(models.Min('privilege'))
-
-        response2 = group_priv['privilege__min']
-        if response2 is None:
-            response2 = PrivilegeCodes.NONE
-        return response2
-
-    def get_effective_user_privilege(self, this_user):
-        """
-        Return the effective user-based privilege of a specific user over this resource
-
-        :param this_user: the user upon which to report
-        :return: integer privilege 1-4 (PrivilegeCodes)
-
-        This accounts for resource flags by revoking CHANGE on immutable resources.
-        """
-        user_priv = self.__get_raw_user_privilege(this_user)
-        if self.immutable and user_priv == PrivilegeCodes.CHANGE:
-            return PrivilegeCodes.VIEW
-        else:
-            return user_priv
-
-    def get_effective_group_privilege(self, this_user):
-        """
-        Return the effective group-based privilege of a specific user over this resource
-
-        :param this_user: the user upon which to report
-        :return: integer privilege 1-4 (PrivilegeCodes)
-
-        This accounts for resource flags by revoking CHANGE on immutable resources.
-        """
-        group_priv = self.__get_raw_group_privilege(this_user)
-        if self.immutable and group_priv == PrivilegeCodes.CHANGE:
-            return PrivilegeCodes.VIEW
-        else:
-            return group_priv
-
-    def get_effective_privilege(self, this_user):
-        """
-        Compute effective privilege of user over a resource, accounting for resource flags.
-
-        :param this_user: user to check.
-        :return: integer privilege 1-4
-
-        This returns the effective privilege of a user over a resource, including both user
-        and group privilege as well as resource flags. Return one of the PrivilegeCodes:
-
-        This overrides stored privileges based upon two resource flags:
-
-            * immutable:
-                privilege is at most VIEW.
-
-            * public:
-                privilege is at least VIEW.
-
-        Recall that *lower* privilege numbers indicate *higher* privilege.
-
-        Usage
-        -----
-
-        This is not normally used in application code.
-        """
-        if __debug__:  # during testing only, check argument types and preconditions
-            assert isinstance(this_user, User)
-
-        if not this_user.is_active:
-            raise PermissionDenied("Grantee user is not active")
-
-        user_priv = self.get_effective_user_privilege(this_user)
-        group_priv = self.get_effective_group_privilege(this_user)
-        return min(user_priv, group_priv)
-
-    @property
-    def sharing_status(self):
-
-        if self.published:
-            return "published"
-        elif self.public:
-            return "public"
-        elif self.discoverable:
-            return "discoverable"
-        else:
-            return "private"
